- 백준 13458 번 문제 totalInspectorCount 를 int 에서 long 으로 변경. - PULL_REQUEST_TEMPLATE.md 작성 완료.

+++ Baekjoon/week_01/Inspector.java
@@ -10,20 +10,20 @@ public class Inspector {
         int N = scan.nextInt();
 
         ArrayList<Integer> A = new ArrayList<>();
-        for (int i = 1; i <= N; i++) {
+        for (int i=1; i<=N; i++) {
             A.add(scan.nextInt());
         }
 
         int B = scan.nextInt(), C = scan.nextInt();
-        int totalInspectorCount = 0;
+        long totalInspectorCount = 0;
 
-        for (int i = 1; i <= N; i++) {
+        for (int i=1; i<=N; i++) {
             int candidateNum = A.get(i - 1);
             candidateNum -= B;
             totalInspectorCount += 1;
 
             if (candidateNum > 0) {
-                totalInspectorCount += (candidateNum + C - 1) / C;
+                    totalInspectorCount += (candidateNum + C - 1) / C;
             }
         }
         System.out.println(totalInspectorCount);
